requirements.txt
pyaudio
openai-whisper
python-docx
keyboard
nltk
sentence-transformers
scikit-learn
pillow
